Another fix for bug 692542 for garbled output with NumRenderingThreads > 1.
The multi-threaded rendering "predicts" which direction to start threads,
but the 'next_band' calculation in clist_get_band_from_thread was not
correct if the thread prediction was incorrect as it would be if the
device requested bands in other than the normal order as happens when
the device makes multiple rendering passes (psdcmyk) or NumCopies > 1.

Also, depending on the number of bands, 'clist_teardown_render_threads'
could free the main thread's 'data' area, then later point the cdev->data
to the freed block, also leaving one of the thread's data blocks leaked
(releasing the chunk memory for the thread would free this, so the leak
would be temporary).