Change the remap_color_info structure so that we can support up to 32 colorants DeviceN
AR supports up to 32 colorants in a DeviceN color space and this brings us in line with
that product's limits.   Note that if the number of colorants is greater than
GS_CLIENT_COLOR_MAX_COMPONENTS then we end up using the alternate tint transform.
Previously, the tint transform would fail if we encountered a color DeviceN color
space with more that GS_CLIENT_COLOR_MAX_COMPONENTS colorants.  This fixes bug 693185