Bug 693216: replace "word-wise" copies with memcpy()
Various places in the blending code used a loop to explicitly copy bytes from
the input buffer to the output, with the comment that the idiom used was
actually quicker than using memcpy(). The idiom in question relies on casting
the byte buffer to a four byte type. This raises the spectre of unaligned
accesses.

Testing suggests any difference between the word-wise copying and "modern"
memcpy implementations is non-existent, or too small to measure (my tests
showed more variation between runs than between the two implementations).

We reckon the memcpy() solution is more maintainable in the long term than
forcing the buffers into alignment.

No cluster differences.