Avoid cmd_put_color being confused by deep color.
With high level color, we can have depths that are too bit to fit in
a gx_color_index. cmd_put_color has no idea how many bits of a
gx_color_index are used, so assumes that as many bits are used as
are in the target depth. The encoding used goes wrong when the number
of bits is bigger it guesses at is bigger than the number of bits in
a gx_color_index.

The fix is simply to change the writing and reading code to be smarter
about guessing; when writing/reading a gx_color_index you can never
need to write more bits than there are actually in a gx_color_index.

While debugging this code, I spotted that the devn_{write,read}_color
routines were always leaving an unused byte at the end of their commands.
Trim that here.