Fix bug 693220. The pdf14 device used compressed encoding in clist mode.
The pdf14 device clist mode did not respect the USE_COMPRESSED_ENCODING
setting always using compressed encoding, and ended up writing pure
colors with num_bytes == -3. This was undetected because tiffsep1 is not
part of the regression testing, and because the tiffsep1 device did not
use 'planar' mode as the tiffsep device did. Also fixed some blanks before
line ends and tab indents.

Also, since planar mode is more efficient and allows for > 8 colorants
tiffsep1 was changed to use planar mode, getting rid of the need for
compressed color encoding in this file (maybe the last one).