Fix clist_copy_planes causing clist corruption.
When clist_copy_planes is used with a planar device, it used to
leave a gap (9 bytes in the debugging example), then write the
first planes data. Then it would write each subsequent plane (preceeded
with a single byte gap). Then it would fill in the 9 bytes.

Unfortunately, if the buffer filled up while copying one of the
non-initial planes in, the date would be copied away and stored before
the header was filled in.

This leads to clist corruption, and all sorts of problems ranging from
rendering issues to SEGVs.

The fix is to move the header writing forward to immediately after
the first plane is successfully written.

This was causing problems with a 72dpi unbanded rendering of
tests_private/comparefiles/Bug692517.pdf to the psdcmyk device
with GS_CLIENT_COLOR_MAX_COMPONENTS set to 32, but was also
causing problems without the change in MAX_COMPONENTS.