Fix link warnings about bad memset in gdevtsep.c
Due to a mistake on my part, the tiffsep device had a couple of
memsets in that did nothing; these were intended to clear an array
of pointers before use. Not clearing the array would only have been
a problem if we'd hit an error condition in a very small region of
code, but nonetheless, this is a fix.