Add common unix style scaling suffix support for -d____= parameters.
Supported suffixes are 'k' and 'K' ( *= 1024), 'm' and 'M' ( *= 1024*1024)
and 'g' and 'G' ( *= 1024*1024*1024). Note that only 1g will not overflow.
No support is provided for .5g or 1.5g values, but this is better than
nothing. Both ghostscript and the other ghostpdl apps (that use plmain.c)
now provide this support.