ps2write - don't leave our dictionary on the stack at temination of the output
Bug #693241 " save / restore around pdf2ps output causes invalidrestore"

Although the output of ps2write isn't intended for encapsulation, we do intend
one day to produce an eps2write (or something) which will be. So we may as
well fix this now.

Previously we left our internal dictionary on the dictionary stack at the
end of the output, here we simply 'end' it which removes it from the stack and
make the output better suited for encapsulation.

No differences expected