Update documentation for tiffsep planar device
Add in comments about the use of -dMaxSpots as well as the fact that the device,
and psdcmyk are planar and limited to 64 colorants per page.  Also add in a
hint about using -dMaxSpots when we are processing a Postscript file and bump
up to the default max limit of 10 colorants.  Tested it and it worked nicely.
Thanks Robin Watts.