pdfwrite - check annotation placement and Box sizes for PDF/X-3 validity
Bug #693233 "PDF/X-3 created using ghostscript gives preflight errors"

We were not checking the bounds of annotations to ensure they were outside
the TrimBox or BleedBox (if present). Annotations are not permitted in the
printable area of a PDF/X-3 file.

In addition we were not checking to see that the TrimBox bounds did not exceed
the BleedBox or MediaBox.

There seems to be a bug in the Acrobat validation tool which insists that the
CropBox must not extend past the BleedBox. I can find no justification for
this in the specification, and indeed the PDF 1.7 spec notes that the CropBox has
'no defined relationship with any of the other boundaries' so for now I've chosen
to ignore this restriction.

No differences expected as PDF/X-3 output is not cluster tested