Workaround libpaper's poor memory management.
libpaper allocates using the stdlib malloc(), and expects the calling code
to release the memory via stdlib free().

Normally this works fine, but Memento hooks malloc() and free() in the
pre-processor, so it sees the free() call, but not the malloc() call.

This can result in confusing, spurious errors from memento.

No cluster differences.