pdfwrite - linearisation.
First attempt to make pdfwrite emit linearised PDF. THis basiclaly works
but needs some attention to the shared object table, and memory cleanup
still needs to be done.

pdfwrite - linearisation project, ma ny changes

Alter rewrite_object to parse the object and alter teh numbers of any objects which it
references

modify rewrite_object to record the offset of the nely rewritten object in the
linearised file, so we cna later write an xref.

Numerous 'off by 1' bug fixes relating to object 0 and whether pages are
numberd from 0 or not.

Write all pages and their objects, and shared objects, in the correct order.

Write a primary xref table.

Record CharProc resource usage by fonts.

pdfwrite - linearisation code

Add code to rewrite the secondary xref with the actual offsets of the objects

Write a dummy primary hint stream

Record the offsets of the primary hint stream and linearisation dictionary.

pdfwrite - linearisation

It turns out that for the hint stream we need to know which objects are used
on every page (shared or not).

This change laters the record for each resoruce to add an array of integers
documenting which page(s) the resource is used on, as well as whether it is
shared or not and so on.

This means that instead of a simple chunk of memory, the record is now a gc
object, properly tracked and relocated and so on.

Added code to track the location of some more objects in the linearised file,
for the benefit of the linearisation dictionary.

Added code to rewrite the linearisation dictionary with the final numbers, and the trailer dict with the correct numbers too.

pdfwrite - linearisation various changes

Remove some unused variables to silence compiler warnings

Sort out the object numbering so that the seocond page is object
number 1 (this is a requirement, but its not obvious in the spec). Make sure
the 'part 9' objects (Info dict, metadata etc) have number that start after
all the page objects, unique objects per page, and shared objects.

Move the hint stream generation so that the hint stream is *after* page 1 and its
resoruces as this is easier to deal with.

pdfwrite - linearisation work

preparing to write hint streams:

first we go back to using a temp file instead of a specific filename (except
when explicitly debugging).

second we only reorder the existing objects and the linearisation dict.

next we copy these re-oredered objects from the temp file, and insert the
dummy primary hint stream at the correct location.

Finally we write the primary xref, but taking the size of the inserted hint stream
into account for the offsets.

In a fit of tidying up, use gp_fseek_64 instead of fseek. Still work to do here.

pdfwrite - linearisation

add preliminary code to write a hint stream (still needs work)

Fix code which wrote teh page object twice for each page.

Various white space tweaking to make the output closer to the MuPDF output
in an abortive attempt to find out why files are not identified as linearised
by Acrobat

pdfwrite - linearisation, make files at least 4k

I don't believe it.....

Acrobat won't recognise a file as linearised unless it is at least 4k!

Also added some comments regarding various implementation notes from the
1.7 reference

pdfwrite - linearisation, move code into a routine of its own

Modularise the linearisation code somewhat by making it a new routine
instead of bloating pdf_close still further.

pdfwrite - linearisation, fixes for hint writing