Fix scaling issues seen with fast halftoning code.
We were missing out the right hand column of source data in images
(see the turkeys on 09-47N page 2 in comparison to png or
USE_FAST_CODE=0 renderings); fix this by not subtracting 1 from
src_size. Don't subtract 1 from data_length (or dest width) either
to match.

Change dda to do a 'half step' to ensure we are sampling from the
centre of each pixel (avoids a shift to the right in zoomed up
images). This requires a new dda init macro - remove the old
(unused, slower) dda formulation to avoid having to think too hard.