Increase max object size to 16M.
Increase the size of rsize member of the ref structure from ushort
to uint_32. This is needed to support large composite objects permitted by current
PDF specification. On 64-bit systems the actual size of ref didn't change.
On 32-bit systems the ref has grown to 12 bytes, which caused a couple of
changes in the alignment and padding code. Finally, the max size of all
composite objects has been increased to 16M.

This patch leaves large objects exposed PS interpreter.
Traditional PS limits can be re-imposed on the PS interpreter if
needed.

The patch results in a couple of progressions on the PDF test base,
and a few expected differences on PS test files that test traditional
limits on composite PS objects.