Fix bug 690963. Update the plane_extract device for hl color handling.
The plane_extract device (used by pksmraw, pamcmyk32, and bmpsep1 as
well as gdevbjc_.c) was never updated for many of the newer device
procedures, including fillpage. This is handled by extending the
prototype procs for the device so that gx_device_forward_fill_in_procs
does not use the procs from the target.