Bug 693279: font: default base encoding
There is some guess work here: Acrobat appears to be always using
StandardEcoding as the default encoding for (at least) some non-symbolic
fonts (rather than, as the spec states, the encoding specified by the font).
As this is contrary to spec, and we only have a CFF example, I've restricted
this to only apply to fonts with Subtype Type1C *and* that result in a Type 2
PS font.

No cluster differences.