Enabling of fast threshold code for color source images as well as monochrome images to cmyk devices
Many cases where we could be using the fast thresholding code were disabled due to issues that this
commit fixes.  Some items such as the conversions to planar when the color management transform was
the identity were not in place.   In addition, for subtractive devices, the polarity of the threshold
array was wrong.