Arrange for gscms_create/destroy to be called correctly.
gscms_create should be called before any other cms call is made,
but it's not being. It not being called has meant that lcms has
always been using malloc/free rather than the gs allocation
functions. Fix some bugs in those allocation functions that show
up now they are used.

Also ensure that pointers are correctly enumerated.

Ensure that cms set context pointer actually makes it into links.

Update gscms interface to ensure that errors on gscms_create can
be responded to.

This work has exposed a problem wrt the context pointer in the
gscms interface, so we fix that here too.

The theory has always been that we can drop new and different
CMS engines into ghostscript, by wrapping each one with some code
that implements the gscms_ functions from gsicc_cms.h.

While this interface was developing, I started to add support for
a context pointer, but looking back at it, it seems I never
completed the work and it was left in a state where it was useless.

After much discussion with Michael, we've decided that the best way
to proceed is to arrange for gscms_create to be called when gs
starts up (just after the base level allocator is created), and for
gscms_destroy to be called as gs exits (just before the base level
allocator is destroyed).

This base level allocator is passed into gscms_create/gscms_destroy
so it can safely allocate memory. Furthermore, we have done away
with explicitly passing the cms_context pointer around, as it was
not really any use in its current form. Instead, we have added
functions to get/set the context pointer in the gs_lib_ctx.

  void *gs_lib_ctx_get_cms_context( const gs_memory_t *mem );
  void gs_lib_ctx_set_cms_context( const gs_memory_t *mem, void *cms_context );

As the gscms functions take mem pointers in all the interesting
places, this should be sufficient.

As mkromfs calls the gs lib ctx init stuff, we have had to add stubs
for gscms_create/destroy to mkromfs.c.