Fix clist_copy_planes to insure that all planes written together. Bug 693234.
The code at 'copy:' checked to make sure that the planes would fit in a
single cbuf, but did not check to make sure that all of the planes fit in
the remaining space in the BufferSpace from 'cnext' to 'cend'. The fix is
to call cmd_write_buffer if all planes don't fit in the current BufferSpace
so that an 'end_run' byte will not be inserted in the data stream.