Add overprint support into the pdf14 device group composing
This fixes Bug 693300.   This this file had two images. One in a separation color space,
the other an RGB color space each in their own transparency group.
Overprint  mode was enabled during the spot image drawing, which occurs after the RGB image is drawn.
During the group pop, there was no code in pdf14_compose_group to account for the overprint condition