Add in option of -dPreBandThreshold=true/false to specify thresholding prior to clist
When halftoning, it may be advantageous from a memory storage to halftone an image pre-clist.
This adds code to check if the halftoned image at device resolution is smaller than the source
image.  It it is, then the image will by default be rendered before going into the clist.
This option is turned off through the use of the device parameter -dPreBandThreshold=false