Simple apitest program.
Requires pthreads. Build with "make apitest" within gs.

Invoke with a ghostscript like command line: e.g.

 bin/apitest -sDEVICE=pdfwrite -o outfile.%d. examples/tiger.eps

The command line must have a -o in it, because it rewrites the
"outfile.%d." to be "outfile.%d.THREADNUM" where THREADNUM is 0 to 9
(change NUM_THREADS in psi/apitest.c to change this).

stdout and stderr are redirected to each stdout.THREADNUM and
stderr.THREADNUM respectively.