pdfwrite - improve subset font prefix generation
We generate the prefix for a subset font by creating a hash from the
usage pattern of glyphs in a font. If the font was already subset, and
'compact' (ie glyphs are used in order from 1 to n), then it was possible
to generate the same prefix for two differetn subsets of the same font.

In fact this is always possible, but here we add a heuristic which uses
the MD5 hash we create for stream objects as an additional hash when
generating the subset prefix. This does not guarantee that name collisions
won't occur but should reduce the incidence still further.

No differences expected.