pdfwrite - track form depth when creating forms with BP pdfmark
Bug #693327

In order to properly set the co-ordinate space for patterns inside forms
we need to know the form's default co-ordinate space. To do this we keep
a record of the nested form depth. However this wasn't being updated when
we created a new form with the /BP pdfmark.

This commit fixes that.