pdfwrite - improve Outline handling when processing a subset of pages from PDF input
The Outline pdfmarks could get confused if a leaf node was elided due to its
Viwe or Dest page lying outside the range of pages processed from a PDF file.
This could lead to nesting going wrong and trying to create more than 32 levels
of nodes.

This commit moves the checking of the Dest and View pages from the PDF interpreter
into pdfwrite, which means pdfwrite now needs to pick up the FirstPage and
LastPage switches. If a destination page lies outside the processed range we
need to preserve it (so that the tree works out properly), but we don't write
the View or Dest property, so that we don't point to non-existent pages.

No differences epxected this is not cluster tested.