Add some PACIFY_VALGRIND magic for helgrind.
Various bits of code (notably gs_debug, some clist function pointers
and some constants in lcms) are held in statics. We accept that
gs_debug is going to be thread unsafe and choose to ignore it.
The others are initialised multiple times (to the same values) and
then never changed.

Helgrind is not smart enough to see that these are repeatedly
updated to the same values, so flags an error every time they are
written to.

We therefore add a bit of Valgrind magic before these places to
tell it to ignore these addresses when checking.