Remove gs_lib_ctx_get_non_gc_memory_t in GS_THREADSAFE builds.
gs_lib_ctx_get_non_gc_memory_t relied on a static to implement it,
so is incompatible with multi-threaded builds. Removing it
causes various devices/bits of code to break. We fix the
important (i.e. mostly non-contrib) devices that use it here.

We also rework the gscms/lcms interface code to avoid it too.

With this commit in, helgrind apitests tiger to ppmraw as being
clean.