pdfwrite - linearisation bug fixes
Allow return code to be 0 when seeking so we don't lose the first part of
the linearised file.

If the input job makes marks on the final page, but does not call showpage,
we discard the final page. This can lead to resources being marked as 'used'
on pages that no longer exist. We now ignore such usage.

Fix an 'off by 1' error when indexing an array using paeg numbers
(page numebrs are 1->n, array is 0->n-1)

Put a #ifdef round some debug prints so they don't end up in the cluster log

Some resources are only used by other resources (eg Encoding) and should take
on the Page Usage of their parent. If the parent resource was shared we werne't
correclty updating the child with all the pages the parent was used on.

Some memory was being double freed causing seg faults.

No differences expetced as cluster doesn't test linearisation.