Bulk of work to demonstrate the use of DeviceN ICC output profile.
This adds support to the psdcmyk device as well as the required changes in
the graphics library.   Through the use of the LIMIT_TO_ICC define in gdevpsd, it is
possible to limit the colorants to those defined by the ICC output profile.
Setting to 1 will limit it setting to 0 will allow all spots (up to the maximum)
to be created.   If spot names in the document match those in the command line
with -sICCOutputColors, then those colorants and hence separation will be treated
the same.