Import LCMS 2.4 (patched)
Here we import LCMS 2.4 with our existing patches into gs.

Cluster testing this shows various changes; most of these are just
small rounding issues. It is possible that some of this is due to
Marti adopting our revised 16 -> 8 code and having pushed it further
through the code than we had.

Also there are changes to the black point compensation code designed
to more accurately match Adobe. For some test files that have
deliberately extreme profiles we get noticably different results.
Supposedly this is because those profiles do not round trip nicely,
but after discussing it with Michael he says it's nothing to worry
about.