Simplify mkromfs build dependencies, add needed debug and error printing.
Since mkromfs minimal memory implementation left the gs_lib_ctx set to NULL,
the errwrite and outwrite functions would de-reference a NULL causing a
Segmentation violation if gp_unix ever printed anything. This removes the
need for the gsmisc.o altogether and removes the need for gslibctx.o and
the dummy gscms functions.