Addition of support for allowing control over the use of black point compensation in color management
Black point compensation is a new extended graphic state entry in the extended graphic
state with PDF 2.0 .  This change adds it to the graphic state.  This update also allows
the same type of override control for black point that we currently have for the
rendering intent.   This includes specification of its use by source object type with
the use of -sSourceObjectICC for source RGB and CMYK graphic, images and text objects.
Alternatively, control can be achieved for the whole document via object type
by specification  -dBlackPtComp=0/1.  Alternatively, you can specify control
on an object based method using  -dGraphBlackPt=0/1
-dImageBlackPtComp=0/1  -dTextBlackPtComp=0/1.  All of these
will require the setting of -dOverrideBP to ensure that the document content does not
interfere with the command line setting.  Note that these are not true/false settings
but integer values like the rendering intent.  This is to allow the easy expansion
to other black point compensation options as described by Bug 692223.