Command line argument character encoding handling rework.
Recap of the situation: Historically gs has always taken command
line args in whatever the 'local' encoding of that OS is; for windows
this means they are encoded as per the currently selected codepage;
for mac/linux this tends to mean utf8.

Similarly, whenever gs goes to access a file, it would pass the filenames
to the platform specific file opening code - on windows this would
assume that it was in the local encoding, and on linux would assume
utf8.

There are downsides to dealing with codepage encoding - not least the
fact that some files can't be accessed easily, or that the behaviour
of PS programs may change according to what platform we are on.

Last year we added code so that gs on windows would run as a unicode app,
and would take the supplied unicode argv and convert to utf8 before
passing them to the core of gs. The file reading functions would then
convert from utf8 back to unicode before opening the files. This
code was disabled by default, but seemed to work well in tests.

1 objection was raised to this - namely that we were redefining the
meaning of the entry points (to take utf8 rather than locally encoded
args) and that this might break legacy code. So, it was suggested that
we add a new entry point specifically for utf8, and that we deprecate
the old entry point.

In starting this work, I spotted another problem; gs supports the
'@file' arg (where more command line args can be given in a file).
The current solution of rewriting the command line args does not
address this.

Instead, we update the gsargs arg handling code with a 'get_codepoint'
function pointer. We update the arg_next function to use this to
decode the supplied args (or OPTIONS or file contents) into simple
unicode characters. We then write these into the returned arg
string as utf8 (or in the case of GS_NO_UTF8 builds, as 8 bit 'clean'
characters).

This commit adds a new gsapi_set_arg_encoding function, that allows
the caller to set the encoding used for the args/strings supplied
to the subsequent functions. By default, 'local' encoding is assumed,
hence preserving existing behaviour. By calling:

 gs_set_arg_encoding(minst, GS_ARG_ENCODING_UTF8);

we can get the UTF8 encoding that we desire. We leave ourselves scope
for supporting other encodings (such as EBCDIC?) in the future.

The commit also changes it so that the gs always runs with UTF8 encoding
internally by default; to avoid it, make with 'GS_NO_UTF8=1'.

For the benefit of windows users we implement UTF16LE encoding (i.e.
wchar_t's). And we expose this through the gsapi level as
gsapi_run_fileW (and gsapi_init_with_argsW) so that .net can bind
nicely to it without any casting trickery being required.

Many thanks to Sags for his tireless efforts and comments while
this work was underway.