pdfwrite - emit partial pages
previously, if an input file made marks on the page but did not call
showpage (or equivalent in other languages), pdfwrite would write all the
objects to the file, but would not add the final partial page to the Pages
tree, which meant that the final page didn't appear.

It seems that Adobe Acrobat behaves differently, and produces the final page
under these conditions. This commit changes our behaviour to match the
observed Adobe behaviour.

This was noticed when running some Quality Logic test files, it appears that
the cluster runs in a mode where these partial pages were always emitted
anyway (because of the CET conformance hacks) so there are only a few files
which display any differences.