Fix DEBUG mode SEGVs; bug 693359
In at least a couple of places I was reading a memory pointer from a
structure that wasn't filled in. Simple fixes here.

This solves the crashing with -ZBb in any file with an image pointed
out by Henry on irc.

I'm sure there will be more such problems due to the wide ranging
nature of the commit that changed this; hopefully all will be as
simple to solve as this.