Reinstate tiffsep1 and svgwrite in the device lists
I accidentally left out tiffsep1 when I consolidated the two lists of TIFF
devices, so putting it back now.

svgwrite was removed a while ago due to being incomplete and flaky, but there
has been ongoing interest in the device, some patches posted, so reinstate it
to make it easier for people to test/debug/develop.

It's included in the "FILES" class of devices.

No cluster differences.