pdfwrite - linearisation completion
Update ps2pdf.htm documentation

CHange parameter name to 'FastWebView'

Don't writ the xref on the initial file, as we don't need it and it makes
the original file slightly smaller, which reduces the chances that we will
have to pad the linearised file. (The linearised file must be no smaller
than the original, because we do not close the file, we just seek around and
rewrite it). Same with the trailer dictionary.

Fix 'rewrite object' so that dictioanry and array objects can be of any size
previously we had assunmed they would be less than 64k, but some complex
files broke that.

If the linearised file is smaller than the original, we need to add white
space to bring it up to the same size, or we will be left with garbage at
the end. Worse, the garbage may well contain a partial invalid xref.

Remove the check for one page files as Adobe specifically allows linearising
single page documents....