Bug #691319 - Preserve Flags (/F) from Text annotations
When creating pdfmarks for the benefit of pdfwrite, the PDF interpreter
only handles two types of annotation, Text and Link. We were handling the
Flags entry for Link annotations, but not for Text annotations.

Here we preserve the Flags for text annotations too.

No differences expected, not tested by cluster.