Add preservation of PageLabels from input PDF files when using pdfwrite
Bug #692901 "Preserve Page labels from input PDF when the output is PDF"

Because hte defined pdfmark /PAGELABEL is not flexible enough to handle all
the possible kinds of page labels in PDF, and because it transpires we don't
emit pdfmarks when producing PostScript with ps2write, this feature does
not use teh pdfmark syntax.

It does work in a broadly similar way, we create a string from the PageLabels
entry in the Catalog dictionary (if present) and then pass that via put_params
with a parameter naem of pdfpagelabels (pdfmark uses a name of pdfmark).

This string is then stored in the Catalog dictionary, and emitted when the
output PDF is created.

I'm unable to test this against the original report, as no specimen PDF file
was provided.

No differences expected, this makes no changes in rendering.