Turn on ability to specify no color management from the srcgtag structure for different objects
With this, it is possible in the srcgtag structure to specify that we will not do color management
for RGB or CMYK graphic, image and text objects.   This is done by specifying the keyword None after
the particular tag.  For example "Graphic_CMYK	None" in the file that is used with the -sSourceObjectICC
command option.  In the above case, all CMYK source graphics would be mapped using the unmanaged color
methods (e.g. 255-X)   Also we added in support for preserveK, override source ICC and black point
compensation on an object based case.    Had to worry about the case where the profile passes through
the clist in the commit and had to add in a new parameter to the profile to ensure that the rendering
conditions are associated with the source profile.  Also fixed a bug to keep from changing the number of
channels in the link format.