Allow use of Device Link ICC Profile for source objects
With the -sSourceObjectICC=filename option, it is now possible to specify a device link source
profile for CMYK/RGB graphic, image and text objects.   When such a profile is specified, the mapping
will go from source directly to device values.  Note that the use of a proofing profile is not
possible and the device profile itself is not used in this transformation.  However the device link
profile that is associated with the device and specified with -sDeviceLinkProfile is included
in the transform.