Bug 693339: Replace malloc() with GS memory allocators in JPX code.
We have 2 implementations of JPX decoding filters based on Luratech
and OpenJpeg libraries. The Luratech one is now clean. The API to
the library supports custom memory allocators.

The OpenJpeg case is more difficult. The calls to malloc() in the
wrapper code has been eliminated, but the library calls malloc()
directly. To use a custom allcator one has to extend the API and
change the signatures of all functions that allocate memory.