pdfwrite - when resizing arrays due to incrementally downloaded fonts, increase CIDToGODMapLength
When TrueType fonts are incrementally downloaded, its possible for new character
codes to exceed the number of glyphs originally defined in the TrueType tables.

When this happens we resize various entries in our copy of the font. However
the change to dynamically storing CIDToGIDMap and CIDSet (for PDF/A) omitted
to increase the variable recording the array size (although the array was correctly
resized).

This led to glyphs with CIDs higher than the initial number of glyphs not having
an entry in CIDToGIDMap, and tehrefore being treated as /.notdef

No differences expected.