Fix multi-threaded rendering use of wrong gs_memory_t.
When we create rendering threads, we create a gs_memory_t * for each.
The idea is that every thread should exclusively use its own gs_memory_t
to avoid conflicts. The gsicc_cache was being setup to use the incorrect
one.

This was found (by Ray) when investigating bug 693361. This solves
this particular error, but does not completely solve the bug.