Second attempt at fixing debugging printing with clists.
The band file reading stream does not have a valid memory pointer
as the structure is allocated on the stack, and to have a valid
memory pointer would make the stream attempt to free the structure
on a close.

Here we add a 'local_memory' pointer that we use both for debugging
printing, and for calls to process_interrupts.