Addition of an internal CMM for direct color replacement of text, graphic and image RGB or CMYK colors
This option is used with the -sSourceObjectICC file.  Similar to the option to set None to request that
no color management be used for a particular object type, here we use the keyword Replace to indicate
that for that object type the colorants will be replaced by the method set up in gsicc_replacecm.c
Currently to demonstrate its usage it provides a "negative" color mapping operation.