Ensure that all devices have non-NULL memory pointers.
We use the memory pointers to get to the lib_ctx, hence having NULL
memory pointers in device structures causes headaches.

Checking the comments in gxdevcli.h it seems that memory should
be NULL iff it's in a static prototype. Therefore anywhere that uses
gx_device_init and passes a NULL pointer is suspect.

Some places call gx_device_init and then set the memory pointer
manually afterwards, to ensure that the reference counting is setup
so that it does not attempt to free the device structure when all
instances are gone. This is to indicate that the device structure
is on the stack.

As a neater alternative to this, we offer a gx_device_init_on_stack
function that does the same thing, but makes this explicit.