Provide way for output pagesize to image size in view___.ps tools.
For viewjpeg.ps and viewgif.ps this adds new procedures "viewJPEGgetsize" and
"viewGIFgetsize" which return the dimensions of the input impage, thus allowing
the caller to set the page size.

For viewmiff.ps, viewpbm.ps, viewpcx.ps they take a new parameter "FITPAGE"
which tell the utilities to fit the page to the image - they still honour the SCALE
parameter, which is also used when setting the pagesize.

CLUSTER_UNTESTED